refactor: break blockfilter circular dependency from bitcoin#27596 work
