+++ src/kernel/chain.cpp
@@ -3,7 +3,6 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <chain.h>
-#include <interfaces/chain.h>
 #include <kernel/chain.h>
 #include <sync.h>
 #include <uint256.h>
